+++ tests/fscripts/demo_return.fs
@@ -1,0 +1,9 @@
+var a = 1;
+if(a == 1) {
+  print("return");
+  return;
+} else {
+  print("end");
+}
+
+